+++ server/src/modules/todo/todoRoutes.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import prisma from '../prismaClient.js';
+import prisma from '../../db/prismaClient.js';
 
 const router = express.Router();
