--- server/utility-app.rest
@@ -1,45 +1,0 @@
-### GET /todo
-GET http://localhost:8080/api/todo
-
-### POST /todo
-POST http://localhost:8080/api/todo
-Content-Type: application/json
-
-{
-  "title": "new to-do item from test"
-}
-
-### UPDATE /todo:id
-PUT http://localhost:8080/api/todo/8
-Content-Type: application/json
-
-{
-  "title": "i updated this todoa"
-}
-
-### DELETE /todo:id
-DELETE http://localhost:8080/api/todo/9
-
-### GET /auth/users
-GET http://localhost:8080/api/auth/users
-
-### POST /auth/register
-POST http://localhost:8080/api/auth/register/
-Content-Type: application/json
-
-{
-  "username": "klein",
-  "password": "123123",
-  "fullname": "klein malikot",
-  "birthdate": "{{$datetime iso8601 '12-08-2023' 1 y}}",
-  "gender": "MALE"
-}
-
-### POST /auth/login
-POST http://localhost:8080/api/auth/login/
-Content-Type: application/json
-
-{
-  "username": "klein",
-  "password": "12312123322"
-}
